requirements.txt
alembic==1.14.0
astroid==3.3.11
black==25.1.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dill==0.4.0
exceptiongroup==1.3.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
inflect==7.5.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
mccabe==0.7.0
more-itertools==10.6.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==3.3.7
pytest==8.4.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
SQLAlchemy==2.0.37
tomli==2.2.1
tomlkit==0.13.3
typeguard==4.4.1
typing_extensions==4.12.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0
Werkzeug==3.1.3
